# expression: make `baseBuiltinFunc` support converting from row-based evaluation to vectorized evaluation