<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
We will implement the vectorized evaluation for all our builtin functions soon.
For each builtin function, we only keep one kind of implementation (row-based or vectorized) since it's hard and unnecessary to maintain these two.

Therefore, we need a converter to make all builtin function support row-based evaluation and vectorized evaluation both since these two kinds of evaluation are needed by users.

This PR makes `baseBuiltinFunc` support converting from row-based evaluation to vectorized evaluation.
### What is changed and how it works?
1. Make baseBuiltinFunc support this converting;
2. Add some unit tests;

#### All changes are new methods or unit tests, so it's easy to review.

After converting row-based evaluation to vectorized evaluation, we get some performance gain, and here is the benchmark:
```
BenchmarkMockDoubleRow/Int-12                     100000             15573 ns/op
BenchmarkMockDoubleRow/Real-12                    100000             15661 ns/op
BenchmarkMockDoubleRow/Decimal-12                  30000             56280 ns/op
BenchmarkMockDoubleRow/Duration-12                100000             18741 ns/op
BenchmarkMockDoubleRow/String-12                   30000             49367 ns/op
BenchmarkMockDoubleRow/Datetime-12                 10000            176552 ns/op
BenchmarkMockDoubleRow/JSON-12                      1000           2222229 ns/op

BenchmarkMockDoubleRow2Vec/Int-12                 200000             10135 ns/op
BenchmarkMockDoubleRow2Vec/Real-12                200000             10105 ns/op
BenchmarkMockDoubleRow2Vec/Decimal-12              30000             47783 ns/op
BenchmarkMockDoubleRow2Vec/Duration-12            100000             14568 ns/op
BenchmarkMockDoubleRow2Vec/String-12               30000             48718 ns/op
BenchmarkMockDoubleRow2Vec/Datetime-12             10000            186369 ns/op
BenchmarkMockDoubleRow2Vec/JSON-12                  1000           2226560 ns/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test