@qw4990 2019-08-13
> The code coverage can be improved.

More tests are added, and now all types are covered by unit tests.