@XuHuaiyu 2019-08-14
/run-common-test
/run-integration-commom-test