@qw4990 2019-08-14
The way to make all builtin functions support this row-to-vec converting is changed.
I remove `self` from `baseBuiltinFunc` since if we use this way, we have to change all builtin functions manually to set the `self` field, which results in a huge change.

Now I add a wrapper builtin function to do this in `getFunction`.

PTAL @zz-jason @XuHuaiyu @coocood